SMS alert implemented, now to host on Heroku and setup 1 minute interval cron jobs

+++ app.py
@@ -2,11 +2,13 @@ from flask import Flask, render_template, url_for, request, redirect
 from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime
 from GMD_PS5 import scrape
+from SMS import send
 
 app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///test.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///GMD.db'
 db = SQLAlchemy(app)
 
+text_alert = "GO GET DAT PLAYSTATION 5 \n\nhttps://www.amazon.com/PlayStation-5-Console/dp/B08FC5L3RG?ref_=ast_sto_dp"
 
 class GMD(db.Model):
     id = db.Column(db.Integer, primary_key=True)
@@ -14,50 +16,50 @@ class GMD(db.Model):
     date_created = db.Column(db.DateTime, default=datetime.utcnow)
 
     def __repr__(self):
-        return '<Task %r>' % self.id
+        return '<Entry %r>' % self.id
 
 
-@app.route('/', methods=['POST', 'GET'])
+@app.route('/', methods=['GET'])
 def index():
-    if request.method == 'POST':
-        task_content = request.form['content']
-        new_task = GMD(content=task_content)
+    entries = GMD.query.order_by(GMD.date_created).limit(30).all()
+    return render_template('index.html', entries=entries)
 
-        try:
-            db.session.add(new_task)
-            db.session.commit()
-            return redirect('/')
-        
-        except:
-            return 'HOUSTON WE HAVE A PROB'
-    else:
-        tasks = GMD.query.order_by(GMD.date_created).limit(30).all()
-        return render_template('index.html', tasks=tasks)
 
 @app.route('/gmd', methods=['GET'])
 def getAvailability():
     data = scrape('https://www.amazon.com/PlayStation-5-Console/dp/B08FC5L3RG?ref_=ast_sto_dp')
     dbLength = len(GMD.query.all())
 
-    if dbLength > 120:
-        oldestEntry = GMD.query.order_by(GMD.date_created).first()
+    if dbLength == 0:
+        newEntry = GMD(content= data["price"] if data["price"] else "Unavailable")
         try:
-            db.session.delete(oldestEntry)
+            db.session.add(newEntry)
             db.session.commit()
+            return redirect('/')
         except:
-            return 'There was a problem deleting that task'
+            return 'There was a problem adding that entry'
 
-    newEntry = GMD(content= data["price"] if data["price"] else "Unavailable")
+    else:
+        if dbLength > 120:
+            oldestEntry = GMD.query.order_by(GMD.date_created).first()
+            try:
+                db.session.delete(oldestEntry)
+                db.session.commit()
+            except:
+                return 'There was a problem deleting that entry'
 
-    try:
-        db.session.add(newEntry)
-        db.session.commit()
-        return redirect('/')
-        
-    except:
-        return 'HOUSTON WE HAVE A PROB'
+        lastEntry = GMD.query.order_by(GMD.date_created.desc()).first()
+        if lastEntry.content == 'Unavailable' and data["price"]:
+            send(text_alert)
 
-   
+        newEntry = GMD(content= data["price"] if data["price"] else "Unavailable")
+        try:
+            db.session.add(newEntry)
+            db.session.commit()
+            return redirect('/')
+        except:
+            return 'There was a problem adding that entry'
+
 
 if __name__ == "__ main__":
     app.run(debug=True)
